fix: use size_t for pos in pd_mux::tick() to match length type
Eliminates signed/unsigned comparison warning when comparing pos
with pdout.dev->length.

+++ src/pdrouting.cpp
@@ -642,7 +642,7 @@ void pdrouting::pd_mux::stop() {
 //! trigger tick
 void pdrouting::pd_mux::tick() {
     try {
-        off_t pos = 0;
+        size_t pos = 0;
 
         if (zero_copy) {
             pdout.dev->push(pdout.provider);
